refactor(risk): 抽离信号词汇表 + 架构重构蓝图
阶段 A 结构解耦(保留全部检测能力与测试,行为逐字节等价):
- 将 SignalID / SignalCategory / RiskContext / RiskScoreReport /
  RiskSignalState 从 Risk/RiskReport.swift 移入新文件
  Risk/RiskSignalVocabulary.swift,使信号类型系统与报告序列化解耦。
  RiskReport.swift 1036 → 744 行;已核验 5 者均为 public 且不引用
  同文件 file-private 类型,括号平衡、无重复定义。
- 新增 CloudPhoneRiskKit_文档/架构重构蓝图.md:现状评估、7 项结构
  性技术债、目标架构、分阶段可编译验证的迁移清单与门禁脚本。

注:当前环境无 Swift 工具链,未做编译验证;合并前请执行 swift test。
