+++ CloudPhoneRiskKit_文档/架构重构蓝图.md
@@ -1,0 +1,120 @@
+# CloudPhoneRiskKit 架构重构蓝图
+
+> 目标:在**完整保留现有检测能力与测试**的前提下,对 7.4 版本的代码结构做分层收敛、模块边界清理与 god-file 解耦。
+> 本文既是现状评估,也是一份**可编译验证的分阶段迁移清单**——每一步都标注了风险与验证方式。
+
+## 0. 环境约束(必读)
+
+本蓝图的代码级迁移**必须在具备 Swift / Xcode 工具链的环境执行**,原因:
+
+- SDK 大量依赖 Apple 专有框架(Security、IOKit、CoreMotion、ObjC runtime),`Package.swift` 只声明 `iOS(.v14)` / `macOS(.v14)`,**无法在 Linux 上编译**。
+- Swift 的 `private` / `fileprivate` 是**文件作用域**。跨文件拆分 god-file 时,若被移动的类型引用了同文件的 file-private 兄弟类型,会在编译期报错。
+- 因此,任何"拆文件"迁移都遵循一条硬规则:
+
+  > **只有顶层 `public` / `internal`(默认)类型,且其函数体不引用同文件的 `private`/`fileprivate` 兄弟,才可零风险整体移动到同一 target 的新文件。**
+  > SwiftPM 按 target 编译,不关心 target 内的目录/文件划分,所以同 target 内移动文件不影响构建。
+
+每个阶段结束都以 `swift test`(SPM)或 `xcodebuild test`(XcodeGen)全绿作为验收门。
+
+## 1. 现状架构评估
+
+### 1.1 模块与分层(`Sources/`)
+
+| Target | 职责 | 规模 |
+|--------|------|------|
+| `CRiskCore`(C) | 原生 VM / 混淆 / 反调试 / 直调 syscall / whitebox 运行时 | ~70 个 `.c/.h` |
+| `CloudPhoneRiskKit`(Swift) | SDK 主体:检测、决策、存储、上报、加固 | 207 个 `.swift` |
+| `CloudPhoneRiskAppCore`(Swift) | 应用层编排(配置加载、检测流程、报告映射) | 8 个 `.swift` |
+
+`CloudPhoneRiskKit` 目录分层(自底向上,**总体健康**):
+
+```
+Util(叶子工具) → Risk(领域内核) → Providers / Detection / Analysis(信号采集)
+   → Decision(判决) → Core(SDK 装配) → 公共入口 CPRiskKit
+横切:Config · Storage · Network · Graph · Behavior · Challenge · TrustChain · Internal/CFF
+```
+
+### 1.2 已确认的结构性技术债
+
+| 编号 | 坏味道 | 事实依据 | 影响 |
+|------|--------|----------|------|
+| D1 | **共享内核类型埋在报告 god-file** | `RiskSignal` / `RiskSignalState` / `SignalID` 原定义于 1036 行的 `Risk/RiskReport.swift`,却被 13+ 目录、32 个文件消费 | 领域内核与序列化耦合,改动放大 |
+| D2 | **CFF 基础设施泄漏进领域层** | `Internal/CFF/CFFDispatcher` 被 `Decision`、`Detection×3`、`TrustChain` 直接 reach-in(7 目录) | 混淆细节污染业务判决逻辑,阻碍 CFF 策略独立演进 |
+| D3 | **DTO 聚合型 god-file** | `Config/RemoteConfig.swift`(34 类型)、`Core/Protocols/ConfigProvider.swift`(36)、`Risk/RiskReport.swift`(26)、`Decision/DecisionTree.swift`(20)、`Analysis/TemporalAnalyzer.swift`(25) | 单文件承载过多类型,难以定位与并行修改 |
+| D4 | **Provider 脚手架重复** | 30 个类型 conform `RiskSignalProvider`,`collect → RiskSignal` 样板重复 | 新增 Provider 成本高,一致性靠人肉保证 |
+| D5 | **加密助手散落** | SHA256/HMAC/CryptoKit 出现在 45 处,`IntegritySealComputer` / `PayloadCrypto` / `StorageIntegrityGuard` / `ConfigSignatureVerifier` / `RiskConclusionSigner` 逻辑重叠 | 密码学正确性分散,审计面大 |
+| D6 | **god-class 单一巨类** | `Detection/Adapter/AntiTamperingSignalProvider.swift`(2737 行,34 private 成员)、`Decision/RiskDetectionEngine.swift`(2318 行) | 单类内聚过载,`private` 密集使跨文件拆分需先收敛访问控制 |
+| D7 | **无构建/CI 入口** | 仓库无 `Makefile`、无 `.github/workflows`、无构建脚本 | 重构缺少自动化回归护栏 |
+
+### 1.3 公共 API 表面(重构须保持不变)
+
+入口类 `CPRiskKit`(`final class: NSObject`),关键 public 方法:`start` / `stop` / `evaluate` / `evaluateAsync` / `buildSecureReportEnvelope*` / `bindAccount` / `updateRemoteConfig` 等;`@objc` 通过 `ObjCBridge/CPRiskKitObjCBridge` 镜像。**任何重构不得改变这些签名与 `@objc` 表面**,否则破坏接入方 ABI(见 `ABI_CONTRACT.md`)。
+
+## 2. 目标架构
+
+在现有分层基础上做三处收敛,不引入新 target(保持 ABI 与打包不变):
+
+1. **确立 Risk 为"领域内核"**:信号类型系统(ID/category/state/RiskSignal 值类型)从报告序列化中分离,成为被上层引用的稳定内核。
+2. **把 CFF 收敛为纯基础设施边界**:领域层不再直接持有 `CFFDispatcher`,改为通过一个窄接口(如 `CFFGate`/协议)注入,`Internal/CFF` 之外只依赖抽象。
+3. **DTO 与逻辑分离**:每个 god-file 拆成"逻辑主体 + 值类型/DTO 附属文件",按类型族归档。
+
+## 3. 分阶段迁移清单
+
+> 每个阶段独立可验收、可回滚。顺序按"风险递增、依赖递进"排列。
+
+### 阶段 A ✅(本次已执行)— 抽离信号词汇表(消 D1 第一步)
+- 将 `SignalID` / `SignalCategory` / `RiskContext` / `RiskScoreReport` / `RiskSignalState` 从 `RiskReport.swift` 移入新文件 `Risk/RiskSignalVocabulary.swift`。
+- 已核验:5 者均为 `public` 且**不引用**同文件 file-private 的 `Payload`/`DetectionResultPayload`;`RiskReport.swift` 1036 → 744 行;括号平衡、无重复定义。
+- 验收门(需 Swift 环境):`swift test` 全绿。
+
+### 阶段 B — RiskSignal 与序列化解耦(消 D1 收尾)
+- `RiskSignal`(public)耦合了 file-private `struct Payload`(见 `RiskReport.swift:378/381`)。需将 `RiskSignal` + `Payload` + `DetectionResultPayload` + `SignalDigest` 作为**一组**移入 `Risk/RiskSignal.swift`;或将 `Payload` 提升为 `internal` 后单独迁移 `RiskSignal`。
+- 风险:中(涉及 file-private 迁移)。验证:全量 Codable round-trip 测试 + `ReportEnvelopeNegativeTests`。
+
+### 阶段 C — DTO god-file 拆分(消 D3)
+- 按"每个类型族一个文件"拆 `RemoteConfig.swift` / `ConfigProvider.swift` / `DecisionTree.swift` / `TemporalAnalyzer.swift`。
+- 迁移前对每个候选类型执行门禁脚本(见 §4),只迁移无 file-private 依赖者;有依赖者先做访问控制收敛。
+- 风险:低-中(逐类型可验证)。验证:对应领域测试族全绿。
+
+### 阶段 D — CFF 基础设施边界化(消 D2)
+- 定义 `CFFGate` 协议(位于 `Internal/CFF`),`RiskDetectionEngine` / `DecisionTree` / `TrustChainManager` / `Detection` 改为依赖协议而非具体 `CFFDispatcher`,由装配层注入。
+- 风险:中-高(触及热路径与 CFF 策略绑定)。验证:`CFFDispatcherTests` / `CFFStateCodecTests` / `CFFRuntimeSaltTests` + `DecisionTreeTests` 全绿,并对比 `cff_policy.yaml` 覆盖建议器输出。
+
+### 阶段 E — god-class 内聚拆分(消 D6)
+- `AntiTamperingSignalProvider`(2737 行)、`RiskDetectionEngine`(2318 行)先在**同文件内**按 `// MARK` + `extension` 分组(零风险),再逐块把不依赖 `private` 成员的部分外移;必要时把跨块共享的 `private` 收敛为 `internal`(注意不扩大 `public` 表面)。
+- 风险:高(private 密集)。验证:`AntiTamperingTests` / `RiskEvaluationActorTests` + 全量 `swift test`。
+
+### 阶段 F — 横切能力抽象(消 D4/D5)
+- D4:抽 `RiskSignalProviderTemplate` 基类/协议扩展,收敛 `collect → RiskSignal` 样板。
+- D5:把 SHA256/HMAC/seal/digest 收敛到 `Util` 下单一 `CryptoPrimitives` 门面,其余站点改为调用。
+- 风险:中。验证:`CryptoTests` / `IntegritySealComputerTests` / `ConfigSignatureVerifierTests` 全绿。
+
+### 阶段 G — 回归护栏(消 D7)
+- 新增最小 CI:`swift test`(SPM)+ `xcodegen && xcodebuild test`;把 `swiftlint` 纳入。使后续阶段有自动回归。
+
+## 4. 迁移门禁脚本(每次拆文件前运行)
+
+对候选迁移类型 `<T>`(所在文件 `<F>`),确认其函数体不引用 `<F>` 内任一 `private`/`fileprivate` 顶层类型:
+
+```bash
+# 1) 列出 <F> 内的 file-private 顶层类型
+grep -nE '^(private|fileprivate) (struct|class|enum|actor|protocol) ' "<F>"
+# 2) 在 <T> 的行范围内检查是否引用了这些类型名;若无输出即可安全迁移
+sed -n '<start>,<end>p' "<F>" | grep -nE '\b(<PrivateTypeA>|<PrivateTypeB>)\b'
+# 3) 迁移后括号平衡自检
+for x in "<F>" "<新文件>"; do echo "$x $(grep -o '{' "$x"|wc -l) $(grep -o '}' "$x"|wc -l)"; done
+```
+
+## 5. 不做什么(边界)
+
+- 不新增/删除 target,不改包产物结构(保持 XCFramework / 静态库打包与 `ABI_CONTRACT.md` 不变)。
+- 不改任何 `public` / `@objc` 方法签名。
+- 不改检测逻辑、阈值、策略语义——纯结构迁移,行为逐字节等价。
+- 不引入第三方依赖。
+
+## 附:本次提交实际变更
+
+- 新增 `Sources/CloudPhoneRiskKit/Risk/RiskSignalVocabulary.swift`。
+- `Sources/CloudPhoneRiskKit/Risk/RiskReport.swift` 移除已迁移的 5 个类型(1036 → 744 行)。
+- 新增本蓝图文档。
+- 由于当前环境无 Swift 工具链,**未做编译验证**;合并前请在 Xcode/Swift 环境执行 `swift test` 确认阶段 A 全绿。
